simplified file structure. still 13 failing tests.
