+++ app/components/lazy-tooltip-on-element.js
@@ -1,0 +1,2 @@
+// TODO can't delete this
+export { default } from 'ember-tooltips/components/lazy-tooltip-on-element';
