--- app/components/popover-on-element-lazy.js
@@ -1,3 +1,0 @@
-export { default } from 'ember-tooltips/components/popover-on-element-lazy';
-
-// TODO(Andrew) find a better name for this
